requirements.txt
asgiref==3.5.2
Django==4.1.1
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.13.1
djongo==1.3.6
dnspython==2.2.1
pymongo==3.12.1
pytz==2022.2.1
sqlparse==0.2.4
